Clarifying phrasing on de.ts
Changed the commaOnlyOnX0 and commaAndOnX0 so the phrasing clearly states that it will be executed EVERY selected Weekday and changed the grammar on commaOnDayX0OfTheMonth since otherwise there would have to be a dot after each stated date eg. "am 14., 16. und 18. des Monats", which does not look good with more than two dates.